Renamed selectall to select-all.
