+++ packages/ckeditor5-select-all/select-all.js
